change field sw to decimal 3,1. + fix seed score subtest 4 ist

+++ migrations/20200715082327-create-score-peserta.js
@@ -15,7 +15,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       sw: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(3, 1)
       },
       kategori: {
         type: Sequelize.STRING
